Oppdater feilmeldingsstate
